server/requirements.txt
Pillow==8.4.0 